app/requirements.txt
Django
python-decouple
psycopg2-binary
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
standardjson==0.3.1
django-prettyjson==0.4.1
